+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added Ruleset and UserFunction to prelude
 - Removed Error types from prelude
+- Moved `parse_json` into the `Rule` type
 
 ## [0.1.1] - 2022-11-22